symbolics: Distinguish user- and internal- Eq
