+++ devito/symbolics/extended_sympy.py
@@ -7,10 +7,9 @@ from sympy import Expr, Float
 from sympy.core.basic import _aresame
 from sympy.functions.elementary.trigonometric import TrigonometricFunction
 
-from devito.region import DOMAIN
 from devito.tools import as_tuple
 
-__all__ = ['FrozenExpr', 'Eq', 'CondEq', 'CondNe', 'Inc', 'Mul', 'Add', 'IntDiv',
+__all__ = ['FrozenExpr', 'Eq', 'CondEq', 'CondNe', 'Mul', 'Add', 'IntDiv',
            'FunctionFromPointer', 'ListInitializer', 'taylor_sin', 'taylor_cos',
            'bhaskara_sin', 'bhaskara_cos']
 
@@ -48,13 +47,9 @@ class Eq(sympy.Eq, FrozenExpr):
 
     """A customized version of :class:`sympy.Eq` which suppresses evaluation."""
 
-    is_Increment = False
-
     def __new__(cls, *args, **kwargs):
         kwargs['evaluate'] = False
-        region = kwargs.pop('region', DOMAIN)
         obj = sympy.Eq.__new__(cls, *args, **kwargs)
-        obj._region = region
         return obj
 
 
@@ -74,15 +69,6 @@ class CondNe(sympy.Ne, FrozenExpr):
     def __new__(cls, *args, **kwargs):
         kwargs['evaluate'] = False
         return sympy.Ne.__new__(cls, *args, **kwargs)
-
-
-class Inc(Eq):
-    """
-    A special :class:`Eq` carrying the information that a linear increment is
-    performed.
-    """
-
-    is_Increment = True
 
 
 class Mul(sympy.Mul, FrozenExpr):
